2025-03-06/jwt-auth/pyproject.toml
[tool.poetry]
name = "jwt-auth"
version = "0.1.0"
description = ""
authors = ["OKnischay <nischaybasi015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
dj-rest-auth = "^7.0.1"
djangorestframework-simplejwt = "^5.5.0"
environs = "^14.1.1"
psycopg = "^3.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
